requirements.txt
requests
Flask
Werkzeug
cryptography
Flask-JWT-Extended
PyJWT
openai
opencv-python-headless
numpy
pillow


